part 4 4.21

+++ part4/blogsList/controllers/blogs.js
@@ -64,14 +64,25 @@ blogsRouter.post('/', async (request, response, next) => {
 
 
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  
   try {
-    const deletedBlog = await Blog.findByIdAndDelete(request.params.id)
-
-    if (deletedBlog) {
-      response.status(204).end() 
-    } else {
-      response.status(404).json({ error: 'Blog not found' }) 
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
     }
+
+    const blog = await Blog.findById(request.params.id);
+    
+    if (!blog) {
+      return response.status(404).json({ error: 'Blog not found' });
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response.status(403).json({ error: 'User not authorized to delete this blog' });
+    }
+
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end(); 
   } catch (error) {
     next(error) 
   }
